+++ src/app/app-routing.module.ts
@@ -1,19 +1,22 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
+import { CharactersPage } from './pages/characters/characters.page';
+import { HomePage } from './pages/home/home.page';
+import { PagesPage } from './pages/pages/pages.page';
 
 const routes: Routes = [
 	{ path: '', redirectTo: 'home', pathMatch: 'full' },
 	{
 		path: 'home',
-		loadChildren: () => import('./pages/home/home.module').then(m => m.HomePageModule)
+		component: HomePage
 	},
 	{
 		path: 'characters',
-		loadChildren: () => import('./pages/characters/characters.module').then(m => m.CharactersPageModule)
+		component: CharactersPage
 	},
 	{
 		path: 'pages',
-		loadChildren: () => import('./pages/pages/pages.module').then(m => m.PagesPageModule)
+		component: PagesPage
 	}
 ];
