--- src/app/pages/characters/characters-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { CharactersPage } from './characters.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: CharactersPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class CharactersPageRoutingModule {}
