+++ src/app/pages/home/home.page.html
@@ -1,13 +1,39 @@
-<ion-header>
-	<ion-toolbar>
-		<ion-title>
-			Home
-		</ion-title>
-	</ion-toolbar>
-</ion-header>
-
 <ion-content [fullscreen]="true">
-	<ion-card (click)="OpenCharactersPage('Members')">
+	<div class="container">
+		<div class="image">
+			<img src="../../../assets/icon/Spire-Round.svg">
+		</div>
+
+
+		<ion-grid>
+			<ion-row>
+				<ion-col size="12" (click)="OpenCharactersPage('Members')">
+					<div>
+						Members
+					</div>
+				</ion-col>
+
+				<ion-col size="12" (click)="OpenCharactersPage('Associates')">
+					<div>
+						Associates
+					</div>
+				</ion-col>
+
+				<ion-col size="12" (click)="OpenCharactersPage('Fallen Heroes')">
+					<div>
+						Fallen Heroes
+					</div>
+				</ion-col>
+
+				<ion-col size="12" *ngFor="let _Category of Categories" (click)="OpenPagesPage(_Category)">
+					<div>
+						{{_Category}}
+					</div>
+				</ion-col>
+			</ion-row>
+		</ion-grid>
+
+		<!-- <ion-card (click)="OpenCharactersPage('Members')">
 		<ion-card-header>
 			<ion-card-title>Members</ion-card-title>
 		</ion-card-header>
@@ -29,5 +55,6 @@
 		<ion-card-header>
 			<ion-card-title>{{_Category}}</ion-card-title>
 		</ion-card-header>
-	</ion-card>
+	</ion-card> -->
+	</div>
 </ion-content>
