+++ src/app/pages/home/home.page.scss
@@ -1,0 +1,30 @@
+.container {
+	width: 100vw;
+	height: 100vh;
+	padding: 10px;
+
+	.image {
+		width: 50%;
+		margin: 0 auto;
+
+		img {
+			width: 100%;
+		}
+	}
+
+	ion-grid {
+		ion-col {
+			--ion-grid-column-padding: 5px;
+			text-align: center;
+
+			div {
+				border: 1px solid var(--ion-color-secondary-shade);
+				padding: 10px;
+				background-color: var(--ion-color-secondary);
+				font-size: 1.5rem;
+				word-wrap: normal;
+				font-weight: bold;
+			}
+		}
+	}
+}
