--- src/app/pages/pages/pages-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { PagesPage } from './pages.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: PagesPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class PagesPageRoutingModule {}
