--- src/app/pages/pages/pages.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { IonicModule } from '@ionic/angular';
-
-import { PagesPageRoutingModule } from './pages-routing.module';
-
-import { PagesPage } from './pages.page';
-import { PipesModule } from 'src/app/pipes/pipes.module';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-	PagesPageRoutingModule,
-	PipesModule
-  ],
-  declarations: [PagesPage]
-})
-export class PagesPageModule {}
